+++ src/js/components/lists/articles/index.jsx
@@ -1,9 +1,15 @@
 import React, { useEffect, useState, Fragment } from 'react';
 import { connect } from 'react-redux';
 import PropType from 'prop-types';
+import get from 'lodash.get';
+
+import FormControl from '@material-ui/core/FormControl';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Grid from '@material-ui/core/Grid';
-import Button from '@material-ui/core/Button';
 import LinearProgress from '@material-ui/core/LinearProgress';
+import Radio from '@material-ui/core/Radio';
+import RadioGroup from '@material-ui/core/RadioGroup';
+import Switch from '@material-ui/core/Switch';
 
 import Card from '../../card';
 import { fetchArticles } from './state/actions';
@@ -17,26 +23,13 @@ const propTypes = {
 };
 
 const articles = ({ articles, articlesLoading, fetchArticles, loadingIndex }) => {
-  const SORT_TITLE = 1;
-  const SORT_CREATE_DATE = 2;
-  const SORT_PUBLISH_DATE = 3;
-  const SORT_TITLE_REVERSE = 4;
-  const SORT_CREATE_DATE_REVERSE = 5;
-  const SORT_PUBLISH_DATE_REVERSE = 6;
-
-  const defaultButtonStyles = { marginRight: '12px' };
-  const selectedButtonStyles = { ...defaultButtonStyles, backgroundColor: 'lightgrey' };
-  const defaultTriangleStyles = { height: '18px', marginLeft: '12px' };
-  const reverseTriangleStyles = { ...defaultTriangleStyles, transform: 'rotate(180deg)' };
-
   const [sortFunction, setSortFunction] = useState(() => _sorts.sortByTitle);
-  const [currentSortOrder, setCurrentSortOrder] = useState(SORT_TITLE);
   const [isLoading, setIsLoading] = useState(false);
 
-  const [titleStyle, setTitleStyle] = useState(selectedButtonStyles);
-  const [createdDateStyle, setCreatedDateStyle] = useState(defaultButtonStyles);
-  const [publishedDateStyle, setPublishedDateStyle] = useState(defaultButtonStyles);
-  const [triangleStyle, setTriangleStyle] = useState(defaultTriangleStyles);
+  const [orderType, setOrderType] = useState('title');
+  const [checkedTitle, setCheckedTitle] = useState(true);
+  const [checkedPublishDate, setCheckedPublishDate] = useState(true);
+  const [checkedCreatedDate, setCheckedCreatedDate] = useState(true);
 
   useEffect(() => {
     fetchArticles();
@@ -46,62 +39,41 @@ const articles = ({ articles, articlesLoading, fetchArticles, loadingIndex }) =>
     setIsLoading(!!loadingIndex || articlesLoading > 0);
   }, [articlesLoading, loadingIndex]);
 
-  const handleSortClick = (sortOrder = -1) => {
-    switch (sortOrder) {
-      case (SORT_TITLE):
-        if (currentSortOrder !== SORT_TITLE) {
-          setSortFunction(() => _sorts.sortByTitle);
-          setCurrentSortOrder(SORT_TITLE);
-          setTriangleStyle(defaultTriangleStyles);
-        } else {
-          setSortFunction(() => _sorts.sortByReverseTitle);
-          setCurrentSortOrder(SORT_TITLE_REVERSE);
-          setTriangleStyle(reverseTriangleStyles);
-        }
-        setTitleStyle(selectedButtonStyles);
-        setCreatedDateStyle(defaultButtonStyles);
-        setPublishedDateStyle(defaultButtonStyles);
-        break;
-      case (SORT_CREATE_DATE):
-        if (currentSortOrder !== SORT_CREATE_DATE) {
-          setSortFunction(() => _sorts.sortByCreatedDate);
-          setCurrentSortOrder(SORT_CREATE_DATE);
-          setTriangleStyle(defaultTriangleStyles);
-        } else {
-          setSortFunction(() => _sorts.sortByReverseCreatedDate);
-          setCurrentSortOrder(SORT_CREATE_DATE_REVERSE);
-          setTriangleStyle(reverseTriangleStyles);
-        }
-        setTitleStyle(defaultButtonStyles);
-        setCreatedDateStyle(selectedButtonStyles);
-        setPublishedDateStyle(defaultButtonStyles);
-        break;
-      case (SORT_PUBLISH_DATE):
-        if (currentSortOrder !== SORT_PUBLISH_DATE) {
-          setSortFunction(() => _sorts.sortByPublishDate);
-          setCurrentSortOrder(SORT_PUBLISH_DATE);
-          setTriangleStyle(defaultTriangleStyles);
-        } else {
-          setSortFunction(() => _sorts.sortByReversePublishDate);
-          setCurrentSortOrder(SORT_PUBLISH_DATE_REVERSE);
-          setTriangleStyle(reverseTriangleStyles);
-        }
-        setTitleStyle(defaultButtonStyles);
-        setCreatedDateStyle(defaultButtonStyles);
-        setPublishedDateStyle(selectedButtonStyles);
-        break;
-      default:
-        setSortFunction(null);
-        setCurrentSortOrder(null);
-        setTitleStyle(defaultButtonStyles);
-        setCreatedDateStyle(defaultButtonStyles);
-        setPublishedDateStyle(defaultButtonStyles);
-        setTriangleStyle(defaultTriangleStyles);
-    };
-  };
+  useEffect(() => {
+    detemineSortOrder();
+  }, [orderType, checkedCreatedDate, checkedPublishDate, checkedTitle]);
 
   const isArticleDisplayable = (article = null) => {
     return !!article && !!article.title && !!((!!article.comments && !!article.comments.length) || (!!article.quotes && !!article.quotes.length));
+  };
+
+  const detemineSortOrder = () => {
+    if (orderType) {
+      let newSort = null;
+      if (orderType === 'title') {
+        newSort = checkedTitle ? _sorts.sortByTitle : _sorts.sortByReverseTitle;
+      } else if (orderType === 'createdDate') {
+        newSort = checkedCreatedDate ? _sorts.sortByCreatedDate : _sorts.sortByReverseCreatedDate;
+      } else if (orderType === 'publishDate') {
+        newSort = checkedPublishDate ? _sorts.sortByPublishDate : _sorts.sortByReversePublishDate;
+      }
+      setSortFunction(() => newSort);
+    }
+  };
+
+  const handleChangeOrderType = (event = {}) => {
+    const newType = get(event, 'target.value', 'title');
+    setOrderType(newType);
+  };
+
+  const handleChangeChecked = (value = '') => {
+    if (value === 'checkedTitle') {
+      setCheckedTitle(!checkedTitle);
+    } else if (value === 'checkedCreatedDate') {
+      setCheckedCreatedDate(!checkedCreatedDate);
+    } else if (value === 'checkedPublishDate') {
+      setCheckedPublishDate(!checkedPublishDate);
+    }
   };
 
   const renderData = () => {
@@ -113,31 +85,61 @@ const articles = ({ articles, articlesLoading, fetchArticles, loadingIndex }) =>
     });
   };
 
+  const getSwitch = (keyName = '', checkName = '', checkValue = false) => {
+    return ((!!keyName) && (!!checkName)) && (
+      <Switch
+        checked={((orderType === keyName) && (checkValue))}
+        color={'primary'}
+        disabled={(orderType !== keyName)}
+        onChange={() => handleChangeChecked(checkName)}
+        value={checkName}
+      />);
+  };
+
   const displayProgressBar = () => {
     return !!isLoading && (<LinearProgress />);
   };
 
-  const sortTriangle = <img src="images/black-sort-triangle.png" style={triangleStyle}></img>;
-
   return (
     <Fragment>
       {displayProgressBar()}
-
       <Grid container spacing={0}>
-        <Grid item style={{ marginTop: '8px' }} sm={12}>
-          <span style={{ fontWeight: 600, marginRight: '12px' }}>{'Order by: '}</span>
-          <Button style={titleStyle} onClick={() => handleSortClick(SORT_TITLE)} size="small">
-            Title
-            {(currentSortOrder === SORT_TITLE || currentSortOrder === SORT_TITLE_REVERSE) && sortTriangle}
-          </Button>
-          <Button style={createdDateStyle} onClick={() => handleSortClick(SORT_CREATE_DATE)} size="small">
-            Created Date
-            {(currentSortOrder === SORT_CREATE_DATE || currentSortOrder === SORT_CREATE_DATE_REVERSE) && sortTriangle}
-          </Button>
-          <Button style={publishedDateStyle} onClick={() => handleSortClick(SORT_PUBLISH_DATE)} size="small">
-            Publish Date
-            {(currentSortOrder === SORT_PUBLISH_DATE || currentSortOrder === SORT_PUBLISH_DATE_REVERSE) && sortTriangle}
-          </Button>
+        <div style={{ fontSize: '24px', fontWeight: '700', textAlign: 'center', width: '100%' }}>Articles List</div>
+        <Grid item style={{ border: '3px solid ' + (isLoading ? '#767676' : '#3f51b5'), borderRadius: '8px', margin: '8px 24px', padding: '12px' }} sm={12}>
+          <div style={{ fontSize: '20px', fontWeight: '700' }}>Sort Order</div>
+          <FormControl>
+            <RadioGroup
+              name={'orderType'}
+              style={{ marginRight: '16px' }}
+              title={'Order Type'}
+              value={orderType}
+              onChange={handleChangeOrderType}
+            >
+              <FormControlLabel control={<Radio color={'primary'} />} disabled={isLoading} label={'Title'} value={'title'}/>
+              <FormControlLabel control={<Radio color={'primary'} />} disabled={isLoading} label={'Created Date'} value={'createdDate'} />
+              <FormControlLabel control={<Radio color={'primary'} />} disabled={isLoading} label={'Publish Date'} value={'publishDate'} />
+            </RadioGroup>
+          </FormControl>
+          <FormControl>
+            <FormControlLabel
+              control={getSwitch('title', 'checkedTitle', checkedTitle)}
+              disabled={isLoading}
+              label={orderType !== 'title' ? '' : checkedTitle ? 'Ascending' : 'Desending'}
+              style={{ padding: '2px' }}
+            />
+            <FormControlLabel
+              control={getSwitch('createdDate', 'checkedCreatedDate', checkedCreatedDate)}
+              disabled={isLoading}
+              label={orderType !== 'createdDate' ? '' : checkedCreatedDate ? 'Ascending' : 'Desending'}
+              style={{ padding: '2px' }}
+            />
+            <FormControlLabel
+              control={getSwitch('publishDate', 'checkedPublishDate', checkedPublishDate)}
+              disabled={isLoading}
+              label={orderType !== 'publishDate' ? '' : checkedPublishDate ? 'Ascending' : 'Desending'}
+              style={{ padding: '2px' }}
+            />
+          </FormControl>
         </Grid>
         {renderData()}
       </Grid>